test works and some repo cleanup
