+++ .gitignore
@@ -1,2 +1,3 @@
 *.sw*
 *.pyc
+.cache/
